Changed f64 to f32

+++ src/geometry/core.rs
@@ -2,13 +2,13 @@
 
 #[derive(Debug)]
 pub struct GeometricHasher {
-    angles: Vec<f64>,
-    distances: Vec<f64>,
+    angles: Vec<f32>,
+    distances: Vec<f32>,
     hash: Vec<u64>, // TODO: Figure out the data type for hash
 }
 
 impl GeometricHasher {
-    pub fn new(angles: Vec<f64>, distances: Vec<f64>) -> GeometricHasher {
+    pub fn new(angles: Vec<f32>, distances: Vec<f32>) -> GeometricHasher {
         GeometricHasher {
             angles: angles,
             distances: distances,
